More playable - op_call fixed, timers and input added
op_call was pushing the current PC (pointing to the call opcode) to the stack
which meant we returned and called the subroutine again in a infinite loop.
Should push the _next_ pc.

timers count down now, but there's no attempt to make them count at 60Hz
or any attempt to control the cpu speed either.

input sometimes works for some keys...

Sound is TBD.